Width-correct individual characters when copying them
Fira Code punctuation characters (a) vary in horizontal advance among
themselves and (b) may not match the horizontal advance or glyph width
of the target font. This change fixes the horizontal advance when
copying individual characters.

Depending on the width of the copied characters (relative to the input
font) and the setting of --scale-character-glyphs-threshold, it will
also either scale the copied glyphs horizontally, or simply center them
inside the now correctly sized hbox.

Note that this only applies to individual characters (and thus is a
no-op if the user hasn't turned on individual character copy with the
--copy-character-glyphs option); properly scaling ligatures is still a
work in progress.